// client/shared/enums/accountPageState.enum.ts
export enum AccountStateEnums {
    SUBSCRIPTION = "subscription",
    ORDER_HISTORY = "orderHistory",
    BILLING_INFO = "billingInfo",
    SHIPPING_INFO = "shippingInfo",
    REVIEWS = "reviews",
    RATED_PRODUCTS = "ratedProducts",
    PERSONAL_INFO = "personalInfo",
    RESET_PASSWORD = "resetPassword"
}